Create filter book by name
closes #21
