+++ work_at_olist/base/tests/test_books_page.py
@@ -4,22 +4,7 @@ from http import HTTPStatus
 import pytest
 from django.urls import reverse
 
-from work_at_olist.base.models import Author, Book
 from work_at_olist.django_assertions import assert_contains, assert_not_contains
-
-
-@pytest.fixture
-def books(db):
-    """
-    Creates and returns some books.
-    """
-    authors = [Author.objects.create(name=f'Author {i}') for i in range(10)]
-    books = []
-    for i, author in enumerate(authors):
-        b = Book.objects.create(name=f'Book {i}', edition=1, publication_year=2024)
-        b.authors.add(author)
-        books.append(b)
-    return books
 
 
 @pytest.fixture
